+++ R/influxdb_database_management.R
@@ -1,7 +1,6 @@
 #' Influx database management
 #'
-#' The folllowing functions are convenient wrappers around `influx_post`
-#' and `influx_query`.
+#' The folllowing functions are convenient wrappers around `influx_post`.
 #' * `create_database()`: creates a new database
 #' * `drop_database()`: drops an existing database
 #' * `drop_series()`: drops specific series
@@ -11,7 +10,6 @@
 #' * `drop_retention_policy()`: drop a retention policy
 #'
 #' @inheritParams influx_query
-#' @param id Sets the series ID.
 #' @param measurement Sets a specific measurement.
 #' @param where Apply filter on tag key values.
 #' @param rp_name The name of the retention policy.
@@ -59,30 +57,23 @@ drop_database <- function(con, db) {
 #' @rdname create_database
 drop_series <- function(con,
                         db,
-                        id = NULL,
                         measurement = NULL,
                         where = NULL) {
+  
   query <- "DROP SERIES"
   
-  if (!is.null(id)) {
-    query <- paste(query, id)
+  query <- ifelse(is.null(measurement),
+                  query,
+                  paste(query, "FROM", measurement))
     
-  } else {
-    query <- ifelse(is.null(measurement),
-                    query,
-                    paste(query, "FROM", measurement))
-    
-    query <- ifelse(is.null(where),
-                    query,
-                    paste(query, "WHERE", where))
-    
-  }
+  query <- ifelse(is.null(where),
+                  query,
+                  paste(query, "WHERE", where))
   
-  result <- influx_query(
+  result <- influx_post(
     con = con,
     db = db,
-    query = query,
-    return_xts = FALSE
+    query = query
   )
   
   if (!is.null(result)) {
@@ -96,12 +87,11 @@ drop_series <- function(con,
 #' @export
 #' @rdname create_database
 drop_measurement <- function(con, db, measurement) {
-  result <- influx_query(
+  result <- influx_post(
     con = con,
     db = db,
     query = paste("DROP MEASUREMENT",
-                  measurement),
-    return_xts = FALSE
+                  measurement)
   )
   
   if (!is.null(result)) {
@@ -115,8 +105,8 @@ drop_measurement <- function(con, db, measurement) {
 #' @export
 #' @rdname create_database
 create_retention_policy <- function(con,
-                                    rp_name,
                                     db,
+                                    rp_name,
                                     duration,
                                     replication,
                                     default = FALSE) {
@@ -136,9 +126,9 @@ create_retention_policy <- function(con,
     
   }
   
-  result <- influx_query(con = con,
-                         query = query,
-                         return_xts = FALSE)
+  result <- influx_post(con = con,
+                        db = db,
+                        query = query)
   
   if (!is.null(result)) {
     return(result)
@@ -151,8 +141,8 @@ create_retention_policy <- function(con,
 #' @export
 #' @rdname create_database
 alter_retention_policy <- function(con,
-                                   rp_name,
                                    db,
+                                   rp_name,
                                    duration,
                                    replication,
                                    default = FALSE) {
@@ -171,9 +161,9 @@ alter_retention_policy <- function(con,
     query <- paste(query, "DEFAULT")
   }
   
-  result <- influx_query(con = con,
-                         query = query,
-                         return_xts = FALSE)
+  result <- influx_post(con = con,
+                         db = db,
+                         query = query)
   
   if (!is.null(result)) {
     return(result)
@@ -185,13 +175,15 @@ alter_retention_policy <- function(con,
 
 #' @export
 #' @rdname create_database
-drop_retention_policy <- function(con, rp_name, db) {
+drop_retention_policy <- function(con,
+                                  db,
+                                  rp_name) {
   query <- paste("DROP RETENTION POLICY", rp_name,
                  "ON", db)
   
-  result <- influx_query(con = con,
-                         query = query,
-                         return_xts = FALSE)
+  result <- influx_post(con = con,
+                        db = db,
+                        query = query)
   
   if (!is.null(result)) {
     return(result)