+++ R/influxdb_json_parser.R
@@ -123,7 +123,6 @@ query_list_to_tibble <- function(x, timestamp_format) {
   ### in case of CHUNKED responses, concatenate tables with same statement_id
   list_of_result <- list_of_result %>% # take the list of results
     `if`(performance, timer(., "concatenate tables with same statement_id"), .) %>%
-    # This may cause trouble in with InfluxDB < V1.2, see issue #32
     purrr::map("statement_id") %>% # extract "statement_id" of each result
     purrr::map_int(unique) %>% # create a vector with unique "statement_id"
     rle %>% # perform run length encoding to get the length of each "statement_id"