+++ NEWS.md
@@ -1,4 +1,4 @@
-# influxdbr 0.13.0.9000
+# influxdbr 0.14.0
 
 ## Breaking changes